requirements.txt
slack_sdk==3.33.4
python-dotenv==1.0.1